Implemented DashAttack and Multiplier/Percentage Attack
On a gamepad press the West ("x") button and you should dash forward, spawning a blue hitbox. When you press the North ("y") button you should spawn a red hitbox. These all have cool downs as well. I have placed a dummy in the game for you test on. ">MULTIPLIER HIT" should print on the debug console when the dummy is hit with the melee attack. ">DASH ATTACK HIT" should print to the debug console when the dummy is hit with a dash attack.